+++ ChangeLog-1.0.md
@@ -4,6 +4,11 @@ All notable changes of the Blast orm 1.0 release series are documented in this f
 
 ## 0.6.0
 
+### Altered
+
+ - add `\Blast\Orm\Entity\DefinitionInterface` and use definition instead of provider
+ - Refactor hydrators use [`zendframework/zend-hydrator`](http://framework.zend.com/manual/current/en/modules/zend.stdlib.hydrator.html) for internal logic.
+
 ### Removed
 
  - remove definition data getters from provider, you need to use `$provider->getDefiniton()->get*()`
