+++ src/Query.php
@@ -17,7 +17,7 @@ use Blast\Orm\Entity\EntityAwareTrait;
 use Blast\Orm\Entity\ProviderFactoryInterface;
 use Blast\Orm\Entity\ProviderFactoryTrait;
 use Blast\Orm\Entity\ProviderInterface;
-use Blast\Orm\Hydrator\ArrayToObjectHydrator;
+use Blast\Orm\Hydrator\Hydrator;
 use Blast\Orm\Hydrator\HydratorInterface;
 use Blast\Orm\Query\Events\QueryBuilderEvent;
 use Blast\Orm\Query\Events\QueryResultEvent;
@@ -150,7 +150,7 @@ class Query implements ConnectionAwareInterface, EmitterAwareInterface,
 
         $result = $this->convertTypesToPHPValues($provider, $event->getResult());
 
-        $data = (new ArrayToObjectHydrator($provider))->hydrate($result, $option);
+        $data = (new Hydrator($provider))->hydrate($result, $option);
         gc_collect_cycles();
 
         return $data;
